main.py

+++ main.py
@@ -16,4 +16,4 @@ print(First == Second)
 
 #4th program
 Number = '123.456'
-print((float(Number) * 10) % 10)
+print(int(float(Number) * 10) % 10)
